+++ ui/src/app.tsx
@@ -36,7 +36,7 @@ import { Storage } from 'pages/project/storage/storage'
 import { Summary } from 'pages/project/summary/summary'
 import { Team } from 'pages/project/team/team'
 import { Projects } from 'pages/projects/projects'
-import SessionDetails from 'pages/session-details/new/session-details'
+import SessionDetails from 'pages/session-details/session-details'
 import { Sessions } from 'pages/sessions/sessions'
 import { Species } from 'pages/species/species'
 import { TaxaListDetails } from 'pages/taxa-list-details/taxa-list-details'
