+++ ui/src/pages/session-details/activity-plot/lazy-activity-plot.tsx
