+++ ui/src/pages/session-details/activity-plot/useDynamicPlotWidth.ts
