+++ ui/src/pages/session-details/capture-info.tsx
