+++ ui/src/pages/session-details/capture-navigation.tsx
