+++ ui/src/pages/session-details/capture/capture.module.scss
