+++ ui/src/pages/session-details/capture/capture.tsx
