+++ ui/src/pages/session-details/hooks/useActiveCapture.ts
