+++ ui/src/pages/session-details/hooks/useActiveOccurrences.ts
