--- ui/src/pages/session-details/playback/activity-plot/lazy-activity-plot.module.scss
@@ -1,6 +1,0 @@
-.loadingWrapper {
-  height: 100px;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
