--- ui/src/pages/session-details/playback/activity-plot/lazy-activity-plot.tsx
@@ -1,21 +1,0 @@
-import { ErrorBoundary } from 'components/error-boundary/error-boundary'
-import { LoadingSpinner } from 'nova-ui-kit'
-import React, { Suspense } from 'react'
-import styles from './lazy-activity-plot.module.scss'
-import { ActivityPlotProps } from './types'
-
-const _ActivityPlot = React.lazy(() => import('./activity-plot'))
-
-export const ActivityPlot = (props: ActivityPlotProps) => (
-  <Suspense
-    fallback={
-      <div className={styles.loadingWrapper}>
-        <LoadingSpinner size={32} />
-      </div>
-    }
-  >
-    <ErrorBoundary>
-      <_ActivityPlot {...props} />
-    </ErrorBoundary>
-  </Suspense>
-)
