--- ui/src/pages/session-details/playback/capture-details/capture-details.module.scss
@@ -1,50 +1,0 @@
-.starButtonWrapper {
-  position: absolute;
-  top: 8px;
-  right: 8px;
-  display: flex;
-  gap: 8px;
-}
-
-.infoWrapper {
-  display: flex;
-  flex-direction: column;
-  gap: 24px;
-  overflow: hidden;
-}
-
-.jobControls {
-  display: flex;
-  flex-wrap: wrap;
-  gap: 16px;
-  margin-top: 4px;
-}
-
-.label {
-  display: block;
-  @include body-overline();
-  font-weight: 600;
-  color: var(--color-neutral-500);
-  margin-bottom: 4px;
-}
-
-.value {
-  display: block;
-  @include body-base();
-  color: var(--color-generic-white);
-
-  &.bubble {
-    height: auto;
-    padding: 6px 8px 4px;
-    border-radius: 4px;
-    border: 1px solid var(--color-neutral-600);
-    background-color: var(--color-neutral-600);
-    font-weight: 600;
-  }
-}
-
-@media only screen and (max-width: $breakpoint-md) {
-  .infoWrapper {
-    display: none;
-  }
-}
