--- ui/src/pages/session-details/playback/capture-navigation/capture-navigation.module.scss
@@ -1,18 +1,0 @@
-.wrapper {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  align-self: flex-start;
-  height: 48px;
-  gap: 32px;
-  padding: 0 8px;
-  border: 1px solid var(--color-neutral-600);
-  border-radius: 24px;
-  box-sizing: border-box;
-
-  span {
-    @include body-overline-small();
-    font-weight: 600;
-    color: var(--color-generic-white);
-  }
-}
