--- ui/src/pages/session-details/playback/frame/frame.module.scss
@@ -1,70 +1,0 @@
-.wrapper {
-  position: relative;
-  width: 100%;
-  height: 0;
-}
-
-.image,
-.overlay,
-.details,
-.detections,
-.loadingWrapper {
-  position: absolute;
-  width: 100%;
-  height: 100%;
-  vertical-align: middle;
-}
-
-.details {
-  .overlay {
-    opacity: 0;
-    transition: opacity 250ms ease-in-out;
-  }
-
-  &.showOverlay {
-    .overlay {
-      opacity: 1;
-    }
-  }
-}
-
-.detection {
-  position: absolute;
-  box-sizing: border-box;
-  outline: 2px solid var(--color-success);
-  border-radius: 2px;
-
-  &.warning {
-    outline-color: var(--color-warning);
-  }
-
-  &.alert {
-    outline-color: var(--color-destructive);
-  }
-
-  &.filtered {
-    outline: 2px solid var(--color-neutral-500);
-  }
-
-  &.active {
-    outline: 2px solid var(--color-secondary-400);
-  }
-
-  &.clickable {
-    &:hover {
-      cursor: pointer;
-    }
-  }
-}
-
-.loadingWrapper {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-
-@media only screen and (max-width: $breakpoint-md) {
-  .wrapper {
-    grid-column: span 2;
-  }
-}
