--- ui/src/pages/session-details/playback/frame/types.ts
@@ -1,6 +1,0 @@
-export interface BoxStyle {
-  width: string
-  height: string
-  top: string
-  left: string
-}
