--- ui/src/pages/session-details/playback/playback.module.scss
@@ -1,103 +1,0 @@
-.wrapper {
-  display: grid;
-  grid-template-columns: 1fr 2fr;
-  border-radius: 8px;
-  overflow: hidden;
-}
-
-.sidebar {
-  background-color: var(--color-neutral-800);
-  border-bottom: 1px solid var(--color-neutral-600);
-  position: relative;
-}
-
-.sidebarContent {
-  width: 100%;
-  height: 100%;
-  position: absolute;
-  top: 0;
-  left: 0;
-  padding: 16px;
-  display: flex;
-  flex-direction: column;
-  gap: 64px;
-  box-sizing: border-box;
-  overflow-y: auto;
-  overflow-x: hidden;
-}
-
-.sidebarSection {
-  display: grid;
-  grid-template-columns: 1fr;
-  gap: 16px;
-
-  .title {
-    display: block;
-    @include body-large();
-    font-weight: 600;
-    color: var(--color-generic-white);
-  }
-
-  .label {
-    display: block;
-    @include body-overline();
-    font-weight: 600;
-    color: var(--color-neutral-500);
-  }
-}
-
-.captureNavigationWrapper {
-  display: flex;
-  align-items: flex-start;
-  margin-bottom: 16px;
-}
-
-.licenseInfoWrapper {
-  flex: 1;
-  text-align: right;
-
-  p {
-    color: var(--color-neutral-500);
-  }
-
-  p a {
-    color: var(--color-generic-white);
-  }
-}
-
-.bottomBar {
-  grid-column: span 2;
-  background-color: var(--color-neutral-800);
-  padding: 32px;
-}
-
-@media only screen and (max-width: $breakpoint-md) {
-  .sidebar {
-    grid-column: span 2;
-  }
-
-  .sidebarContent {
-    position: static;
-  }
-
-  .sidebarSection:last-child {
-    display: none;
-  }
-
-  .bottomBar {
-    display: flex;
-    flex-direction: column;
-    padding: 16px;
-  }
-
-  .captureNavigationWrapper {
-    margin-bottom: 32px;
-    gap: 32px;
-    flex-direction: column-reverse;
-    justify-content: flex-start;
-  }
-
-  .licenseInfoWrapper {
-    text-align: left;
-  }
-}
