+++ ui/src/pages/session-details/process/process-now.tsx
