--- ui/src/pages/session-details/session-details.module.scss
@@ -1,47 +1,0 @@
-.loadingWrapper {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  min-height: 320px;
-}
-
-.main {
-  position: relative;
-}
-
-.fetchInfoWrapper {
-  position: absolute;
-  height: 50px;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  top: 0;
-  right: 0;
-  transform: translateY(-100%);
-}
-
-.playbackWrapper {
-  display: flex;
-  align-items: stretch;
-  justify-content: space-between;
-  margin-bottom: 32px;
-
-  > * {
-    width: 100%;
-  }
-}
-
-.sessionInfo {
-  width: 100%;
-}
-
-.text {
-  @include body-base();
-  color: var(--color-foreground);
-}
-
-@media only screen and (max-width: $breakpoint-md) {
-  .playbackWrapper {
-    margin-bottom: 16px;
-  }
-}
