+++ ui/src/pages/session-details/session-info.tsx
