--- ui/src/pages/session-details/session-info/session-info.module.scss
@@ -1,13 +1,0 @@
-.content {
-  width: 100%;
-  display: grid;
-  grid-template-columns: 1fr 1fr;
-  column-gap: 16px;
-}
-
-.title {
-  @include body-large();
-  font-weight: 600;
-  color: var(--color-foreground);
-  margin: 0 0 16px 0;
-}
