--- ui/src/pages/session-details/session-info/session-info.tsx
@@ -1,71 +1,0 @@
-import { Session } from 'data-services/models/session'
-import { InfoBlock } from 'nova-ui-kit'
-import { useParams } from 'react-router-dom'
-import { APP_ROUTES } from 'utils/constants'
-import { getAppRoute } from 'utils/getAppRoute'
-import { STRING, translate } from 'utils/language'
-import styles from './session-info.module.scss'
-
-export const SessionInfo = ({ session }: { session: Session }) => {
-  const { projectId } = useParams()
-
-  const fields = [
-    {
-      label: translate(STRING.FIELD_LABEL_ID),
-      value: session.id,
-    },
-    {
-      label: translate(STRING.FIELD_LABEL_DEPLOYMENT),
-      value: session.deploymentLabel,
-      to: APP_ROUTES.DEPLOYMENT_DETAILS({
-        projectId: projectId as string,
-        deploymentId: session.deploymentId,
-      }),
-    },
-    {
-      label: translate(STRING.FIELD_LABEL_DATE),
-      value: session.datespanLabel,
-    },
-    {
-      label: translate(STRING.FIELD_LABEL_TIME),
-      value: session.timespanLabel,
-    },
-    {
-      label: translate(STRING.FIELD_LABEL_DURATION),
-      value: session.durationLabel,
-    },
-    {
-      label: translate(STRING.FIELD_LABEL_CAPTURES),
-      value: session.numImages,
-    },
-    {
-      label: translate(STRING.FIELD_LABEL_OCCURRENCES),
-      value: session.numOccurrences,
-      to: getAppRoute({
-        to: APP_ROUTES.OCCURRENCES({ projectId: projectId as string }),
-        filters: { event: session.id },
-      }),
-    },
-    ...(session.numTaxa !== undefined
-      ? [
-          {
-            label: translate(STRING.FIELD_LABEL_TAXA),
-            value: session.numTaxa,
-            to: getAppRoute({
-              to: APP_ROUTES.TAXA({ projectId: projectId as string }),
-              filters: { event: session.id },
-            }),
-          },
-        ]
-      : []),
-  ]
-
-  return (
-    <div className={styles.container}>
-      <h1 className={styles.title}>{session.label}</h1>
-      <div className={styles.content}>
-        <InfoBlock fields={fields} />
-      </div>
-    </div>
-  )
-}
