+++ ui/src/pages/session-details/session-plots.tsx
