+++ ui/src/pages/session-details/star-button.tsx
