+++ ui/src/pages/session-details/timeline-slider/styles.module.scss
