+++ ui/src/pages/session-details/utils.tsx
