+++ ui/src/pages/session-details/view-settings.tsx
